Move SQL normalization to the sql module.
